py: introduce common for async/sync commonalities
also make the example mqtt client use unit for exit request event
