+++ py/miniconf-mqtt/miniconf/__main__.py
@@ -1,9 +1,5 @@
-"""
-Command line interface for Miniconf-over-MQTT (asynchronous)
-"""
-
-import asyncio
-from .async_ import _main
-
 if __name__ == "__main__":
+    import asyncio
+    from .async_ import _main
+
     asyncio.run(_main())
