+++ py/miniconf-mqtt/miniconf/async_.py
@@ -6,7 +6,6 @@ Asynchronous Miniconf-over-MQTT utilities
 
 import asyncio
 import json
-import logging
 import uuid
 from typing import Dict, Any
 
@@ -14,20 +13,7 @@ import paho.mqtt
 from paho.mqtt.properties import Properties, PacketTypes
 from aiomqtt import Client, Message, MqttError
 
-MQTTv5 = paho.mqtt.enums.MQTTProtocolVersion.MQTTv5
-
-LOGGER = logging.getLogger(__name__)
-
-
-class MiniconfException(Exception):
-    """Miniconf Error"""
-
-    def __init__(self, code, message):
-        self.code = code
-        self.message = message
-
-    def __repr__(self):
-        return f"{self.code}: {self.message}"
+from .common import MiniconfException, LOGGER
 
 
 class Miniconf:
@@ -63,7 +49,7 @@ class Miniconf:
 
     async def _listen(self):
         await self.client.subscribe(self.response_topic)
-        LOGGER.info(f"Subscribed to {self.response_topic}")
+        LOGGER.debug(f"Subscribed to {self.response_topic}")
         self.subscribed.set()
         try:
             async for message in self.client.messages:
@@ -76,7 +62,7 @@ class Miniconf:
             try:
                 await self.client.unsubscribe(self.response_topic)
                 self.subscribed.clear()
-                LOGGER.info(f"Unsubscribed from {self.response_topic}")
+                LOGGER.debug(f"Unsubscribed from {self.response_topic}")
             except MqttError as e:
                 LOGGER.debug(f"MQTT Error {e}", exc_info=True)
 
@@ -136,7 +122,7 @@ class Miniconf:
             self._inflight[cd] = fut, []
 
         topic = f"{self.prefix}/settings{path}"
-        LOGGER.info(f"Publishing {topic}: {kwargs.get('payload')}, [{props}]")
+        LOGGER.debug(f"Publishing {topic}: {kwargs.get('payload')}, [{props}]")
         await self.client.publish(
             topic,
             properties=props,
@@ -244,9 +230,9 @@ async def discover(
             try:
                 payload = json.loads(message.payload)
             except json.JSONDecodeError:
-                logging.info(f"Ignoring {peer} not/invalid alive")
+                LOGGER.info(f"Ignoring {peer} not/invalid alive")
             else:
-                logging.info(f"Discovered {peer} alive")
+                LOGGER.debug(f"Discovered {peer} alive")
                 discovered[peer] = payload
 
     t_start = asyncio.get_running_loop().time()
@@ -264,87 +250,11 @@ async def discover(
     return discovered
 
 
-def one(devices: Dict[str, Any]) -> (str, Any):
-    """Return the prefix for the unique alive Miniconf device.
-
-    See `discover()` for arguments.
-    """
-    try:
-        (device,) = devices.items()
-    except ValueError:
-        raise MiniconfException(
-            "Discover", f"No unique Miniconf device (found `{devices}`)."
-        )
-    logging.info("Found device: %s", device)
-    return device
-
-
-class _Path:
-    def __init__(self):
-        self.current = ""
-
-    def normalize(self, path):
-        """Return an absolute normalized path and update current absolute reference."""
-        if path.startswith("/") or not path:
-            self.current = path[: path.rfind("/")]
-        else:
-            path = f"{self.current}/{path}"
-        assert path.startswith("/") or not path
-        return path
-
-
-def _cli():
-    import argparse
-
-    parser = argparse.ArgumentParser(
-        description="Miniconf command line interface.",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-        epilog="""Examples (with a target at prefix 'app/id' and device-discovery):
-%(prog)s -d app/+ /path       # GET
-%(prog)s -d app/+ /path=value # SET
-%(prog)s -d app/+ /path=      # CLEAR
-%(prog)s -d app/+ /path?      # LIST-GET
-%(prog)s -d app/+ /path!      # DUMP
-""",
-    )
-    parser.add_argument(
-        "-v", "--verbose", action="count", default=0, help="Increase logging verbosity"
-    )
-    parser.add_argument(
-        "--broker", "-b", default="mqtt", type=str, help="The MQTT broker address"
-    )
-    parser.add_argument(
-        "--retain",
-        "-r",
-        default=False,
-        action="store_true",
-        help="Retain the settings that are being set on the broker",
-    )
-    parser.add_argument(
-        "--discover", "-d", action="store_true", help="Detect device prefix"
-    )
-    parser.add_argument(
-        "prefix",
-        type=str,
-        help="The MQTT topic prefix of the target or a prefix filter for discovery",
-    )
-    parser.add_argument(
-        "commands",
-        metavar="CMD",
-        nargs="*",
-        help="Path to get ('PATH') or path and JSON encoded value to set "
-        "('PATH=VALUE') or path to clear ('PATH=') or path to list ('PATH?') or "
-        "path to dump ('PATH!'). "
-        "Use sufficient shell quoting/escaping. "
-        "Absolute PATHs are empty or start with a '/'. "
-        "All other PATHs are relative to the last absolute PATH.",
-    )
-    return parser
-
-
 async def _main():
     import sys
     import os
+    import logging
+    from .common import _cli, MQTTv5, one
 
     if sys.platform.lower() == "win32" or os.name.lower() == "nt":
         from asyncio import set_event_loop_policy, WindowsSelectorEventLoopPolicy
@@ -358,9 +268,7 @@ async def _main():
         level=logging.WARN - 10 * args.verbose,
     )
 
-    async with Client(
-        args.broker, protocol=MQTTv5, logger=logging.getLogger("aiomqtt-client")
-    ) as client:
+    async with Client(args.broker, protocol=MQTTv5) as client:
         if args.discover:
             prefix, _alive = one(await discover(client, args.prefix))
         else:
@@ -376,6 +284,7 @@ async def _main():
 
 async def _handle_commands(interface, commands, retain):
     import sys
+    from .common import _Path
 
     current = _Path()
     for arg in commands:
